Deleting Post using jwt

+++ controllers/api/v1/posts_api.js
@@ -30,18 +30,19 @@ module.exports.destroy= async function(req,res){
         //Checking if current users are same
         // console.log(req.params.id);
         
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.remove();
          await Comment.deleteMany({post:req.params.id});
 
          return res.json(200,{
              message:"Post and associated comments deleted successfully"
          })
-        // }else
-        // {   
-        //     req.flash('error', 'Action cannot be performed')
-        //     return res.redirect('back')
-        // }
+        }else
+        {   
+            return res.json(401,{
+                message:"You cannot delete this post"
+            });
+        }
     }catch(err)
     { console.log('****',err)
           return res.json(500,{
